move error handling to separate file
